Add border, debugging tools, and fix teleport
Windows now refresh a border as a mechanism in refresh gaurd.
windo border is defined in the draw_border method of the window and
screen class. I also added debugging tools for slowing down the speed of
the game forcefully by increaseing sleep time in intervals of 4ms with
the minus key. I also fixed a bug in teleport see the TODO that was
there about 'literal corner cases'.